+++ src/main/java/com/github/xburning/dubboswitch/view/consumer/ZookeeperConsumerManageUI.java
@@ -3,6 +3,7 @@ package com.github.xburning.dubboswitch.view.consumer;
 
 import com.github.xburning.dubboswitch.entity.ZookeeperConsumer;
 import com.github.xburning.dubboswitch.repository.ZookeeperConsumerRepository;
+import com.vaadin.data.Item;
 import com.vaadin.server.FontAwesome;
 import com.vaadin.spring.annotation.SpringUI;
 import com.vaadin.ui.*;
@@ -65,8 +66,11 @@ public class ZookeeperConsumerManageUI extends VerticalLayout{
             }
             //batch delete
             for (Object object : items) {
-                Long id = (Long) grid.getContainerDataSource().getItem(object).getItemProperty("序号").getValue();
-                zookeeperConsumerRepository.delete(id);
+                Item item = grid.getContainerDataSource().getItem(object);
+                if (item != null) {
+                    Long id = (Long) item.getItemProperty("序号").getValue();
+                    zookeeperConsumerRepository.delete(id);
+                }
             }
             search();
         });
